handle_read_auto: 0xFF is a valid base value's bottom nibble
When reading autorecorded values and encountering an 0xFF byte, only
consider it invalid if we are not currently waiting for the bottom nibble
of a new base value. For example, the byte sequence 0xF4 0xFF is valid
and encodes a new base value 0x4F (79). However, 0xFF not preceded by a
byte 0xFn (where n != F) still signifies an invalid value.